feat!: make result mutable again on useFacet
